requirements.txt
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.4
flask==2.3.3
python-dotenv==1.0.0
sentence-transformers==2.2.2
anthropic==0.17.0
langchain-anthropic==0.1.2